consolidate conflict resolution logic
